+++ embedded-hal-async/src/delay.rs
@@ -1,6 +1,6 @@
-//! Delays
+//! Delays.
 
-/// Microsecond delay
+/// Microsecond delay.
 pub trait DelayUs {
     /// Pauses execution for at minimum `us` microseconds. Pause can be longer
     /// if the implementation requires it due to precision/timing issues.
@@ -15,10 +15,12 @@ impl<T> DelayUs for &mut T
 where
     T: DelayUs + ?Sized,
 {
+    #[inline]
     async fn delay_us(&mut self, us: u32) {
         T::delay_us(self, us).await
     }
 
+    #[inline]
     async fn delay_ms(&mut self, ms: u32) {
         T::delay_ms(self, ms).await
     }