+++ embedded-hal/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Minor document fixes.
+- Add #[inline] hints to most of `embedded-hal` functions.
+
 ## [v1.0.0-rc.1] - 2023-08-15
 
 - The Minimum Supported Rust Version (MSRV) is now 1.60.0